+++ visualization/app/app.config.ts
@@ -22,6 +22,8 @@ import { UpdateQueryParametersEffect } from "app/codeCharta/state/effects/update
 import { UpdateVisibleTopLabelsEffect } from "app/codeCharta/state/effects/updateVisibleTopLabels/updateVisibleTopLabels.effect"
 import { ResetColorRangeEffect } from "app/codeCharta/state/store/dynamicSettings/colorRange/resetColorRange.effect"
 import { BlacklistSearchPatternEffect } from "app/codeCharta/ui/ribbonBar/searchPanel/searchBar/blacklistSearchPattern.effect"
+import { UpdateShowLabelsEffect } from "./codeCharta/state/effects/updateShowLabels/updateShowLabels.effect"
+import { UpdateAmountOfEdgePreviewsEffect } from "./codeCharta/state/effects/amountOfEdgePreviews/updateAmountOfEdgePreviews.effect"
 
 export const appConfig: ApplicationConfig = {
     providers: [
@@ -33,6 +35,7 @@ export const appConfig: ApplicationConfig = {
         provideEffects([
             UnfocusNodesEffect,
             AddBlacklistItemsIfNotResultsInEmptyMapEffect,
+            UpdateAmountOfEdgePreviewsEffect,
             OpenNodeContextMenuEffect,
             BlacklistSearchPatternEffect,
             ResetColorRangeEffect,
@@ -41,6 +44,7 @@ export const appConfig: ApplicationConfig = {
             RenderCodeMapEffect,
             AutoFitCodeMapEffect,
             UpdateVisibleTopLabelsEffect,
+            UpdateShowLabelsEffect,
             LinkColorMetricToHeightMetricEffect,
             ResetSelectedEdgeMetricWhenItDoesntExistAnymoreEffect,
             UpdateFileSettingsEffect,