+++ visualization/app/codeCharta/ui/ribbonBar/edgeSettingsPanel/edgeSettingsPanel.component.spec.ts
@@ -1,45 +1,51 @@
 import { TestBed } from "@angular/core/testing"
 import { render, screen } from "@testing-library/angular"
 import { EdgeSettingsPanelComponent } from "./edgeSettingsPanel.component"
-import { provideMockStore } from "@ngrx/store/testing"
-import { amountOfBuildingsWithSelectedEdgeMetricSelector } from "./selectors/amountOfBuildingsWithSelectedEdgeMetric.selector"
-import { amountOfEdgePreviewsSelector } from "../../../state/store/appSettings/amountOfEdgePreviews/amountOfEdgePreviews.selector"
-import { edgeHeightSelector } from "../../../state/store/appSettings/edgeHeight/edgeHeight.selector"
-import { showOnlyBuildingsWithEdgesSelector } from "../../../state/store/appSettings/showOnlyBuildingsWithEdges/showOnlyBuildingsWithEdges.selector"
-import { State } from "@ngrx/store"
-import { colorMetricSelector } from "../../../state/store/dynamicSettings/colorMetric/colorMetric.selector"
-import { colorRangeSelector } from "../../../state/store/dynamicSettings/colorRange/colorRange.selector"
-import { mapColorsSelector } from "../../../state/store/appSettings/mapColors/mapColors.selector"
-import { defaultMapColors } from "../../../state/store/appSettings/mapColors/mapColors.reducer"
+import { Store, StoreModule } from "@ngrx/store"
+import userEvent from "@testing-library/user-event"
+import { setState } from "../../../state/store/state.actions"
+import { appReducers, setStateMiddleware } from "../../../state/store/state.manager"
 
 describe("EdgeSettingsPanelComponent", () => {
-    beforeEach(() => {
+    beforeEach(async () => {
         TestBed.configureTestingModule({
-            imports: [EdgeSettingsPanelComponent],
-            providers: [
-                provideMockStore({
-                    selectors: [
-                        { selector: amountOfBuildingsWithSelectedEdgeMetricSelector, value: 10 },
-                        { selector: amountOfEdgePreviewsSelector, value: 10 },
-                        { selector: edgeHeightSelector, value: 10 },
-                        { selector: showOnlyBuildingsWithEdgesSelector, value: true },
-                        { selector: colorMetricSelector, value: "rloc" },
-                        { selector: colorRangeSelector, value: { from: 21, to: 42, max: 9001 } },
-                        { selector: mapColorsSelector, value: defaultMapColors }
-                    ]
-                }),
-                { provide: State, useValue: {} }
-            ]
+            imports: [EdgeSettingsPanelComponent, StoreModule.forRoot(appReducers, { metaReducers: [setStateMiddleware] })]
         })
+
+        await render(EdgeSettingsPanelComponent)
     })
 
-    it("should render correctly", async () => {
-        await render(EdgeSettingsPanelComponent)
+    it("should render correctly", () => {
         expect(screen.getByText("Preview")).toBeTruthy()
         expect(screen.getByText("Height")).toBeTruthy()
         expect(screen.getByText("Outgoing Edge")).toBeTruthy()
         expect(screen.getByText("Incoming Edge")).toBeTruthy()
+        expect(screen.getAllByText("Show")).toHaveLength(2)
+        expect(screen.getAllByRole("checkbox", { name: "Show", checked: true })).toHaveLength(2)
         expect(screen.getByText("Only show nodes with edges")).toBeTruthy()
         expect(screen.getByTitle("Reset edge metric settings to their defaults")).toBeTruthy()
+    })
+
+    it("should dispatch all possible settings when reset button is clicked", async () => {
+        const dispatchSpy = jest.spyOn(TestBed.inject(Store), "dispatch")
+        const resetSettingsButton = screen.getByTitle("Reset edge metric settings to their defaults")
+        const defaultSettings = {
+            appSettings: {
+                amountOfEdgePreviews: 1,
+                edgeHeight: 4,
+                mapColors: {
+                    outgoingEdge: "#ff00ff",
+                    incomingEdge: "#00ffff"
+                },
+                showOutgoingEdges: true,
+                showIncomingEdges: true,
+                showOnlyBuildingsWithEdges: false,
+                isEdgeMetricVisible: true
+            }
+        }
+
+        await userEvent.click(resetSettingsButton)
+
+        expect(dispatchSpy).toHaveBeenCalledWith(setState({ value: defaultSettings }))
     })
 })